+++ src/main/java/com/github/theredbrain/bettercombatextension/BetterCombatExtensionClient.java
@@ -1,12 +1,15 @@
 package com.github.theredbrain.bettercombatextension;
 
+import com.github.theredbrain.bettercombatextension.config.ClientConfig;
 import com.github.theredbrain.bettercombatextension.network.packet.CancelAttackPacket;
 import com.github.theredbrain.bettercombatextension.network.packet.CancelAttackPacketReceiver;
-import com.github.theredbrain.bettercombatextension.network.packet.ServerConfigSyncPacket;
+import me.fzzyhmstrs.fzzy_config.api.ConfigApiJava;
+import me.fzzyhmstrs.fzzy_config.api.RegisterType;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.networking.v1.ClientPlayNetworking;
 
 public class BetterCombatExtensionClient implements ClientModInitializer {
+	public static ClientConfig CLIENT_CONFIG = ConfigApiJava.registerAndLoadConfig(ClientConfig::new, RegisterType.CLIENT);
 
 	@Override
 	public void onInitializeClient() {
@@ -14,8 +17,5 @@ public class BetterCombatExtensionClient implements ClientModInitializer {
 		// Packets
 		ClientPlayNetworking.registerGlobalReceiver(CancelAttackPacket.PACKET_ID, new CancelAttackPacketReceiver());
 
-		ClientPlayNetworking.registerGlobalReceiver(ServerConfigSyncPacket.PACKET_ID, (payload, context) -> {
-			BetterCombatExtension.serverConfig = payload.serverConfig();
-		});
 	}
 }