+++ src/main/java/com/github/theredbrain/bettercombatextension/config/ServerConfig.java
@@ -1,24 +1,28 @@
 package com.github.theredbrain.bettercombatextension.config;
 
-import me.shedaniel.autoconfig.ConfigData;
-import me.shedaniel.autoconfig.annotation.Config;
-import me.shedaniel.cloth.clothconfig.shadowed.blue.endless.jankson.Comment;
+import com.github.theredbrain.bettercombatextension.BetterCombatExtension;
+import me.fzzyhmstrs.fzzy_config.annotations.Comment;
+import me.fzzyhmstrs.fzzy_config.annotations.ConvertFrom;
+import me.fzzyhmstrs.fzzy_config.config.Config;
+import me.fzzyhmstrs.fzzy_config.validation.misc.ValidatedBoolean;
+import me.fzzyhmstrs.fzzy_config.validation.number.ValidatedFloat;
 
-@Config(
-		name = "server"
-)
-public class ServerConfig implements ConfigData {
+@ConvertFrom(fileName = "server.json5", folder = "bettercombatextension")
+public class ServerConfig extends Config {
+	public ServerConfig() {
+		super(BetterCombatExtension.identifier("server"));
+	}
 	@Comment("""
 			WARNING: EXPERIMENTAL
 			
-			Better Combats attacking with weapons in the offhand slot is incompatible with RPG Inventory
+			In 1.21.1, Better Combats attacking with weapons in the offhand slot is incompatible with RPG Inventory
 			
 			Setting this option to true, fixes the incompatibility. This may have unexpected consequences.
 			Please report any issues you encounter when using this setting
 			
 			Default: false
 			""")
-	public boolean enable_experimental_swap_hand_attributes_algorithm = false;
+	public ValidatedBoolean enable_experimental_swap_hand_attributes_algorithm = new ValidatedBoolean(false);
 	@Comment("""
 			When set to true, changing the players orientation and position while attacking is disabled.
 			It is recommended to install Shoulder Surfing Reloaded, play in the third person perspective and activate the decoupled camera setting.
@@ -27,27 +31,27 @@ public class ServerConfig implements ConfigData {
 			
 			Default: false
 			""")
-	public boolean enable_movement_locking_attacks = false;
+	public ValidatedBoolean enable_movement_locking_attacks = new ValidatedBoolean(false);
 	@Comment("""
 			When set to true, jumping is disabled during attacks.
 			Use the item tag "disables_jump_restriction_during_attack" and the entity type tag "disables_jump_restriction_when_ridden" to control when the jump restriction is applied.
 			
 			Default: false
 			""")
-	public boolean enable_jump_restriction_during_attacks = false;
+	public ValidatedBoolean enable_jump_restriction_during_attacks = new ValidatedBoolean(false);
 	@Comment("""
 			When set to true, the pitch of the players attack is restricted.
 			
 			Default: false
 			""")
-	public boolean restrict_attack_pitch = false;
+	public ValidatedBoolean restrict_attack_pitch = new ValidatedBoolean(false);
 	@Comment("""
 			Only has an effect when 'restrict_attack_pitch' is set to 'true'.
 			Describes the angle from 0, which means the final range is from -'attack_pitch_range' to 'attack_pitch_range'.
 			
 			Default: 15.0
 			""")
-	public float attack_pitch_range = 15.0F;
+	public ValidatedFloat attack_pitch_range = new ValidatedFloat(15.0F);
 	@Comment("""
 			Enables an alternative two_handed condition.
 			If the offhand stack is empty and this is true, the mainhand stack is two_handed
@@ -55,20 +59,19 @@ public class ServerConfig implements ConfigData {
 			
 			Default: false
 			""")
-	public boolean empty_offhand_equals_two_handing_mainhand = false;
+	public ValidatedBoolean empty_offhand_equals_two_handing_mainhand = new ValidatedBoolean(false);
 	@Comment("""
-			Disables Better Combat's formerly client feature of
-			continuously attacking while holding down the attack key.
+			Disables Better Combat's formerly client feature of continuously attacking while holding down the attack key.
 			
 			Default: false
 			""")
-	public boolean disable_better_combat_hold_to_attack = false;
+	public ValidatedBoolean disable_better_combat_hold_to_attack = new ValidatedBoolean(false);
 	@Comment("""
 			When set to true, feinting an attack skips it.
 			
 			Default: false
 			""")
-	public boolean feinting_increases_combo_count = false;
+	public ValidatedBoolean feinting_increases_combo_count = new ValidatedBoolean(false);
 	@Comment("""
 			When set to true, the "minecraft:player.entity_interaction_range" entity attribute is used to determine the attack range.
 			It is recommended to set the "isTooltipAttackRangeEnabled" value in the Better Combat client config to false,
@@ -76,13 +79,7 @@ public class ServerConfig implements ConfigData {
 			
 			Default: false
 			""")
-	public boolean use_entity_interaction_range_attribute_as_attack_range = false;
-	@Comment("""
-			When set to false, weapon poses are not displayed while the player is sprinting.
-			
-			Default: true
-			""")
-	public boolean enable_poses_while_sprinting = true;
+	public ValidatedBoolean use_entity_interaction_range_attribute_as_attack_range = new ValidatedBoolean(false);
 	@Comment("""
 			
 			
@@ -90,15 +87,11 @@ public class ServerConfig implements ConfigData {
 			
 			Default: 1.0
 			""")
-	public float global_feint_stamina_cost_multiplier = 1.0F;
+	public ValidatedFloat global_feint_stamina_cost_multiplier = new ValidatedFloat(1.0F);
 	@Comment("""
 			The stamina cost of all attacks is multiplied with this value when the attack is executed.
 			
 			Default: 1.0
 			""")
-	public float global_attack_stamina_cost_multiplier = 1.0F;
-
-	public ServerConfig() {
-
-	}
+	public ValidatedFloat global_attack_stamina_cost_multiplier = new ValidatedFloat(1.0F);
 }